+++ tests/conftest.py
@@ -6,7 +6,7 @@ import sqlalchemy as sa
 from sqlalchemy import create_engine, text
 from sqlalchemy.orm import close_all_sessions, declarative_base, sessionmaker
 
-from postgresql_audit import VersioningManager
+from postgresql_audit import AuditLogger
 
 
 @pytest.fixture
@@ -62,21 +62,21 @@ def schema_name():
 
 
 @pytest.fixture
-def versioning_manager(base, schema_name):
-    vm = VersioningManager(schema_name=schema_name)
-    vm.init(base)
-    yield vm
-    vm.remove_listeners()
+def audit_logger(base, schema_name):
+    al = AuditLogger(schema_name=schema_name)
+    al.init(base)
+    yield al
+    al.remove_listeners()
 
 
 @pytest.fixture
-def activity_cls(versioning_manager):
-    return versioning_manager.activity_cls
+def activity_cls(audit_logger):
+    return audit_logger.activity_cls
 
 
 @pytest.fixture
-def transaction_cls(versioning_manager):
-    return versioning_manager.transaction_cls
+def transaction_cls(audit_logger):
+    return audit_logger.transaction_cls
 
 
 @pytest.fixture
@@ -112,15 +112,15 @@ def models(user_class, article_class):
 @pytest.fixture
 def table_creator(
     base,
-    versioning_manager,
+    audit_logger,
     engine,
     models
 ):
     sa.orm.configure_mappers()
 
     with engine.begin() as connection:
-        versioning_manager.transaction_cls.__table__.create(connection)
-        versioning_manager.activity_cls.__table__.create(connection)
+        audit_logger.transaction_cls.__table__.create(connection)
+        audit_logger.activity_cls.__table__.create(connection)
         base.metadata.create_all(connection)
 
     yield